backend/requirements.txt
fastapi
uvicorn
firebase-admin
pillow
opencv-python
numpy
scikit-learn
python-multipart
